requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dj-database-url==2.1.0
Django==4.2.11
django-autoslug==1.9.9
django-environ==0.11.2
gunicorn==22.0.0
h11==0.14.0
idna==3.7
kombu==5.3.7
packaging==24.0
Pillow==10.0.1
prompt-toolkit==3.0.43
psycopg2==2.9.9
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
stripe==9.4.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
